+++ pmd-core/src/main/java/net/sourceforge/pmd/lang/rule/RuleSetWriter.java
@@ -271,9 +271,11 @@ public class RuleSetWriter {
             // For each provided PropertyDescriptor
 
             PropertyTypeId typeId = InternalApiBridge.getTypeId(descriptor);
+            // RuleReferences can't define additional properties
+            boolean isPropertyDefinition = typeId != null && !(propertySource instanceof RuleReference);
 
-            if (typeId == null // not defined externally
-                && !overridden.contains(descriptor)) {
+            // skip properties, which neither are definitions nor override the default value
+            if (!isPropertyDefinition && !overridden.contains(descriptor)) {
                 continue;
             }
 
@@ -281,10 +283,10 @@ public class RuleSetWriter {
                 propertiesElement = createPropertiesElement();
             }
 
-            if (typeId != null) {
+            if (isPropertyDefinition) {
                 propertiesElement.appendChild(createPropertyDefinitionElementBR(descriptor, typeId));
             } else {
-                propertiesElement.appendChild(propertyElementWithValue(propertySource, descriptor));
+                propertiesElement.appendChild(propertyElementWithValueAttribute(propertySource, descriptor));
             }
         }
 
@@ -292,8 +294,14 @@ public class RuleSetWriter {
     }
 
     @NonNull
-    private <T> Element propertyElementWithValue(PropertySource propertySource, PropertyDescriptor<T> descriptor) {
-        return createPropertyValueElement(descriptor, propertySource.getProperty(descriptor));
+    private <T> Element propertyElementWithValueAttribute(PropertySource propertySource, PropertyDescriptor<T> propertyDescriptor) {
+        Element element = document.createElementNS(RULESET_2_0_0_NS_URI, "property");
+        SchemaConstants.NAME.setOn(element, propertyDescriptor.name());
+
+        PropertySerializer<T> xmlStrategy = propertyDescriptor.serializer();
+        T value = propertySource.getProperty(propertyDescriptor);
+        SchemaConstants.PROPERTY_VALUE.setOn(element, xmlStrategy.toString(value));
+        return element;
     }
 
     private <T> Element createPropertyValueElement(PropertyDescriptor<T> propertyDescriptor, T value) {